kernelci.build: add Step._run_make() method
Add the Step._run_make() method to use instead of the _run_make()
function, so it can be changed and use the Step object attributes
without breaking the existing build_kernel() implementation.  This
uses a separate _get_make_opts() method to just gather all the options
to pass to "make".

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>